Move finalizeConnection back to JetpackConnectionAction
